requirements.txt
aiogram==3.4.1
aiosqlite==0.20.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.3.0
pylibdmtx==0.1.10
pdfplumber==0.11.4
gspread==6.1.0
google-auth==2.29.0

